refactor: simplify FLB ServiceReconciler
